move convertor to a dedicated directory
