+++ src/libs/bpmn-js/MainGraph.ts
@@ -1,7 +1,7 @@
 import { mxgraph } from 'mxgraph';
 import AbstractGraph, { EVENT_WIDTH, EVENT_Y_LITTLE, LANE_HEIGHT_LITTLE, LANE_WIDTH, TASK_HEIGHT, TASK_Y_LARGE, TASK_Y_LITTLE } from './AbstractGraph';
 import BpmnProcessCreatorExampleCodeOnly from './model/BpmnProcessCreatorExample';
-import ModelConvertor from './model/BpmnModelToMxGraphModel';
+import ModelConvertor from './convertor/BpmnModelToMxGraphModel';
 
 export default class MainGraph extends AbstractGraph {
   constructor(container: Element) {
